+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGES
 
+Unreleased
+* Switch markdown parsers to [flexmark-java](https://github.com/vsch/flexmark-java) to address [#199](https://github.com/clj-commons/marginalia/issues/199).
+* Support docstrings in `defmethod`.
+* Clean up the parsing tests.
+* Remove non-functioning css for `<tr>`.
+* Clean up `source-excluded` to do less work.
+* Update tools.cli to 1.4.256.
+* Update tools.namespace to 1.5.1.
+
 0.9.2 -- 01-31-2024
 * Add `:meta` support to address [#186](https://github.com/clj-commons/marginalia/issues/186).
 * Update all dependencies.